finished json layout
