+++ LuckParser/Models/JsonModels/JsonBuffDamageModifierData.cs
@@ -6,39 +6,45 @@ using System.Threading.Tasks;
 
 namespace LuckParser.Models.JsonModels
 {
-    /// <summary>
-    /// Class corresponding to a buff based damage modifier
-    /// </summary>
     public class JsonBuffDamageModifierData
     {
         /// <summary>
-        /// Hits done under the buff
+        /// Class corresponding to a buff based damage modifier
         /// </summary>
-        public int HitCount;
-        /// <summary>
-        /// Total hits
-        /// </summary>
-        public int TotalHitCount;
-        /// <summary>
-        /// Gained damage
-        /// </summary>
-        public int DamageGain;
-        /// <summary>
-        /// Total damage done
-        /// </summary>
-        public int TotalDamage;
-        /// <summary>
-        /// True if the buff is a multiplier type, in which case a % gain makes sense
-        /// </summary>
-        public bool Multiplier;
-
-        public JsonBuffDamageModifierData(ParseModels.AbstractMasterActor.ExtraBoonData extraData)
+        public class JsonBuffDamageModifierItem
         {
-            HitCount = extraData.HitCount;
-            TotalHitCount = extraData.TotalHitCount;
-            DamageGain = extraData.DamageGain;
-            TotalDamage = extraData.TotalDamage;
-            Multiplier = extraData.Multiplier;
+            /// <summary>
+            /// Hits done under the buff
+            /// </summary>
+            public int HitCount;
+            /// <summary>
+            /// Total hits
+            /// </summary>
+            public int TotalHitCount;
+            /// <summary>
+            /// Gained damage
+            /// </summary>
+            public int DamageGain;
+            /// <summary>
+            /// Total damage done
+            /// </summary>
+            public int TotalDamage;
+            /// <summary>
+            /// True if the buff is a multiplier type, in which case a % gain makes sense
+            /// </summary>
+            public bool Multiplier;
+
+            public JsonBuffDamageModifierItem(ParseModels.AbstractMasterActor.ExtraBoonData extraData)
+            {
+                HitCount = extraData.HitCount;
+                TotalHitCount = extraData.TotalHitCount;
+                DamageGain = extraData.DamageGain;
+                TotalDamage = extraData.TotalDamage;
+                Multiplier = extraData.Multiplier;
+            }
         }
+
+        public long ID;
+        public List<JsonBuffDamageModifierItem> DamageModifiers;
     }
 }
