+++ LuckParser/Models/JsonModels/JsonBuffs.cs
@@ -9,6 +9,30 @@ namespace LuckParser.Models.JsonModels
 {
     public class JsonBuffs
     {
+        public class JsonBuffsData
+        {
+            public double Uptime;
+            public double Generation;
+            public double Overstack;
+            public double Wasted;
+            public double UnknownExtension;
+            public double Extension;
+            public double Extended;
+            public double Presence;
+
+            public JsonBuffsData(Statistics.FinalBuffs stats)
+            {
+                Uptime = stats.Uptime;
+                Generation = stats.Generation;
+                Overstack = stats.Overstack;
+                Wasted = stats.Wasted;
+                UnknownExtension = stats.UnknownExtension;
+                Extended = stats.Extended;
+                Presence = stats.Presence;
+            }
+
+        }
+
         /// <summary>
         /// ID of the buff
         /// </summary>
@@ -19,7 +43,7 @@ namespace LuckParser.Models.JsonModels
         /// Length == # of phases
         /// </summary>
         /// <seealso cref="JsonBuffsData"/>
-        public List<JsonBuffsData> Data;
+        public List<JsonBuffsData> BuffData;
         /// <summary>
         /// Array of int[2] that represents the number of the given buff status \n
         /// Value[i][0] will be the time, value[i][1] will be the number of the buff present from value[i][0] to value[i+1][0] \n
