+++ LuckParser/Models/JsonModels/JsonMechanics.cs
@@ -12,6 +12,21 @@ namespace LuckParser.Models.JsonModels
     public class JsonMechanics
     {
         /// <summary>
+        /// Class corresponding to a mechanic event
+        /// </summary>
+        public class JsonMechanic
+        {
+            /// <summary>
+            /// Time a which the event happened
+            /// </summary>
+            public long Time;
+            /// <summary>
+            /// The actor who was hit by the mechanic
+            /// </summary>
+            public string Actor;
+        }
+
+        /// <summary>
         /// List of mechanics application
         /// </summary>
         public List<JsonMechanic> MechanicsData;
